# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Pets-project-app-backend",
    "description": "This is the API of the pets application."
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/pets/": {
      "post": {
        "tags": ["Pets"],
        "summary": "Add a new pet",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 16
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "breed": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 16
                  },
                  "comments": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 120
                  }
                },
                "required": ["name", "date", "breed", "comments"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new pet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 16
                    },
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "breed": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 16
                    },
                    "comments": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 120
                    },
                    "favorite": {
                      "type": "boolean"
                    },
                    "avatarURL": {
                      "type": "string"
                    },
                    "owner": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Pets"],
        "summary": "Get a list of pets",
        "operationId": "listPets",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "default": 1
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "breed": {
                    "type": "string"
                  },
                  "comments": {
                    "type": "string"
                  },
                  "favorite": {
                    "type": "boolean"
                  },
                  "avatarURL": {
                    "type": "string"
                  },
                  "owner": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/pets/{petId}": {
      "delete": {
        "tags": ["Pets"],
        "summary": "Remove pet by ID",
        "description": "Remove a pet by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "petId",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "pet deleted"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Pets"],
        "operationId": "getPetById",
        "summary": "Get pet by id",
        "description": "Retrieve information about a pet based on its id",
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "description": "The id of the pet",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "required": true
                    },
                    "date": {
                      "type": "string",
                      "format": "date-time",
                      "required": true
                    },
                    "breed": {
                      "type": "string",
                      "required": true
                    },
                    "comments": {
                      "type": "string",
                      "required": true
                    },
                    "favorite": {
                      "type": "boolean",
                      "default": false
                    },
                    "avatarURL": {
                      "type": "string",
                      "default": null
                    },
                    "owner": {
                      "type": "string",
                      "required": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/pets/{petId}/update": {
      "put": {
        "tags": ["Pets"],
        "summary": "Update a pet",
        "description": "Updates a pet's information by its id",
        "parameters": [
          {
            "in": "path",
            "name": "petId",
            "description": "The id of the pet to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "The new information for the pet",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 16
                },
                "date": {
                  "type": "string",
                  "format": "date-time"
                },
                "breed": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 16
                },
                "comments": {
                  "type": "string",
                  "minLength": 8,
                  "maxLength": 120
                },
                "favorite": {
                  "type": "boolean"
                },
                "avatarURL": {
                  "type": "string"
                },
                "owner": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pet updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string",
                      "required": true
                    },
                    "date": {
                      "type": "string",
                      "format": "date-time",
                      "required": true
                    },
                    "breed": {
                      "type": "string",
                      "required": true
                    },
                    "comments": {
                      "type": "string",
                      "required": true
                    },
                    "favorite": {
                      "type": "boolean",
                      "default": false
                    },
                    "avatarURL": {
                      "type": "string",
                      "default": null
                    },
                    "owner": {
                      "type": "string",
                      "required": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/pets/avatars/{petId}": {
      "patch": {
        "tags": ["Pets"],
        "summary": "Update pet avatar",
        "parameters": [
          {
            "in": "path",
            "name": "petId",
            "description": "Pet ID",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "formData",
            "name": "avatar",
            "description": "Avatar image",
            "required": true,
            "type": "file"
          }
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "avatarURL": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Notice not found"
          },
          "500": {
            "description": "Server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT authorization token"
    }
  },
...
}